Pattern_Recognition/assignment/requirements.txt
numpy
pip
pandas
matplotlib
Keras
h5py
pillow
scikit-learn
scipy
tensorflow-gpu
joblib
